requirements.txt
Flask
python-dotenv
folium
geopy
requests
numpy
pandas
pytest
black
flake8
colorlog
